[AsyncResult] Migrate Invoke method to managed
This is invoked for every BeginInvoke in the runtime. By switching it to managed, we reduce the numer of LOC in the runtime, leading to less possible very hard to track bugs. We also reduce the number of transitions from managed to native and from native to managed, leading to better performance. Another added benefit is exception handling is better managed.