[WebConnection] Do not change WebConnection queue order on Abort
In case we would abort a HttpWebRequest, we would abort its corresponding WebConnection. Unfortunately, the WebConnection.Abort code to remove this HttpWebRequest from the WebConnection.queue would be wrong: it would reenqueue the element in the opposite order of which they were dequeued. This would lead to requests which have been created after, being executed before.

This is possibly due to the specifics of the Queue implementation, which have been imported from referencesource.